Fix fatal time_diff impl
The previous impl may accidentally add one second when `end` is crossing to next second, e.g. diff of sec field is 1 second, and ns (nanosecond) of `end` is 0.2s, ns for `start` is 0.6s, we will obtain -0.2s of diff, here, I mistakenly add one second to the negative result, which should be add to tv_sec field naturally.